Zenflows.VF: Use decimal in DB, schema and resolvers

+++ src/zenflows/vf/duration/type.ex
@@ -28,7 +28,7 @@ A number representing the duration, will be paired with a unit.
 @desc "A `Duration` represents an interval between two `DateTime` values."
 object :duration do
 	@desc @numeric_duration
-	field :numeric_duration, non_null(:float)
+	field :numeric_duration, non_null(:decimal)
 
 	@desc @unit_type
 	field :unit_type, non_null(:time_unit)
@@ -37,7 +37,7 @@ end
 @desc "Mutation input structure for defining time durations."
 input_object :iduration, name: "IDuration" do
 	@desc @numeric_duration
-	field :numeric_duration, non_null(:float)
+	field :numeric_duration, non_null(:decimal)
 
 	@desc @unit_type
 	field :unit_type, non_null(:time_unit)
