+++ src/zenflows/vf/satisfaction.ex
@@ -45,10 +45,10 @@ schema "vf_satisfaction" do
 	belongs_to :satisfies, Intent
 	field :resource_quantity, :map, virtual: true
 	belongs_to :resource_quantity_has_unit, Unit
-	field :resource_quantity_has_numerical_value, :float
+	field :resource_quantity_has_numerical_value, :decimal
 	field :effort_quantity, :map, virtual: true
 	belongs_to :effort_quantity_has_unit, Unit
-	field :effort_quantity_has_numerical_value, :float
+	field :effort_quantity_has_numerical_value, :decimal
 	field :note, :string
 	timestamps()
 end
